Restructure code into binary and lib crates
To benchmark this it made more sense (or was necessary?) for the important code to be in the structure of a library crate.

I created two modules, `original` and `refactored`, that contain the two versions of the code. `main.rs` then pulls that code in and does some simple calls just to prove that it all still works.