Use try-map to simplify error checking
I really wanted to use `?` on the different applications of `u8::try_from`, but you can't do that _inside_ a closure because the `?` will just get you out of the closure and not the whole function.

Some searching turned up `try_map`, which will do exactly what I want. Unfortunately it's currently only in the nightly build. Hmph. So I've put it on a branch for now, and I'll have to think about what to do about it later.